update test-che-default to be run as an actor
