+++ test/e2e/test-che-default/actor/apify.json
@@ -1,0 +1,6 @@
+{
+	"name": "test-che-default",
+	"version": "0.0",
+	"buildTag": "latest",
+	"env": null
+}
