+++ test/e2e/test-che-default/actor/package.json
@@ -1,0 +1,21 @@
+{
+	"name": "test-che-default",
+	"version": "0.0.1",
+	"description": "Test - Cheerio Crawler Default",
+	"dependencies": {
+        "apify": "../../../../packages/apify",
+        "@crawlee/cheerio": "../../../../packages/cheerio-crawler"
+	},
+	"devDependencies": {
+		"@apify/eslint-config": "^0.3.1",
+		"eslint": "^8.13.0"
+	},
+	"scripts": {
+		"start": "node main.js",
+		"lint": "./node_modules/.bin/eslint ./src --ext .js,.jsx",
+		"lint:fix": "./node_modules/.bin/eslint ./src --ext .js,.jsx --fix",
+		"test": "echo \"Error: oops, the actor has no tests yet, sad!\" && exit 1"
+	},
+	"author": "It's not you it's me",
+	"license": "ISC"
+}
